REFACTOR: create game to test case
