+++ tests/whist/server/api/game/test_join.py
@@ -1,12 +1,6 @@
-from tests.whist.server.api.game.base_token_case import TestCaseWithToken
+from tests.whist.server.api.game.base_created_case import BaseCreateGameTestCase
 
-
-class JoinGameTestCase(TestCaseWithToken):
-    def setUp(self) -> None:
-        super().setUp()
-        self.game_creds = {'game_name': 'test', 'password': 'abc'}
-        response = self.client.post(url='/game/create', json=self.game_creds, headers=self.headers)
-        self.game_id = response.json()['game_id']
+class JoinGameTestCase(BaseCreateGameTestCase):
 
     def test_join(self):
         headers = self.create_and_auth_user('miles', 'abc')
